requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
sqlalchemy==2.0.23
asyncio==3.4.3
networkx==3.2.1
structlog==24.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
docker==7.0.0
pytest==7.4.3
python-dotenv==1.0.0
sqlalchemy-utils==0.41.1
aiohttp==3.9.0
openai==1.3.7
python-dotenv==1.0.0